renaming parameters -> constants
