+++ src/kem/kyber/arithmetic.rs
@@ -1,6 +1,6 @@
 use std::ops::{self, Index, IndexMut};
 
-use super::parameters::{COEFFICIENTS_IN_RING_ELEMENT, FIELD_MODULUS};
+use super::constants::{COEFFICIENTS_IN_RING_ELEMENT, FIELD_MODULUS};
 
 pub(crate) type KyberFieldElement = i32;
